Move server setup from cmd/fleet to new pkg/server
